Prototype
The prototype is made out of oak wood, hinges, eye hooks, steel rods,
and string. We tested the prototype by tying a total of twenty-five
pounds in sand bags onto the chair. By pulling the strings which are
tied onto the eye hooks the top of the chair moved up, bending the
hinges. The prototype took a total of three days to assemble.